code 31.03.2024

+++ database/migrations/2024_02_04_171805_create_sizes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('sizes', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('factor')->default(1);
+            $table->double('factor')->default(1);
             $table->timestamps();
         });
     }
